+++ tests/testthat/test_arith.R
@@ -195,11 +195,11 @@ test_that("%/% and %% work", {
   expect_error(drop_units(set_units(x, foot^2)) %% y)
 })
 
-test_that("%*% work", {
+#test_that("%*% work", {
   # mat = set_units(matrix(1:5, 1), m)
   # expect_equal(mat %*% t(mat), set_units(matrix(55), m^2))
   # expect_equal(t(mat) %*% mat, set_units(t(unclass(mat)) %*% unclass(mat), m^2))
-})
+#})
 
 test_that("The order of multiplication for basic units is commutative", {
   a <- set_units(1:4, m^-3)
@@ -250,3 +250,23 @@ test_that("inverse units can always be multiplied and return unitless (related t
   )
 })
 
+test_that("identical unit multiplication and division respect 'simplify' option (#355)", {
+  initial_unit_simplify <- units_options("simplify")
+  on.exit(units_options(simplify=initial_unit_simplify))
+
+  gg <- structure(list(numerator="g", denominator="g"), class="symbolic_units")
+  x_gg <- set_units(1, gg, mode="standard")
+  x_ul <- set_units(1, unitless, mode="standard")
+
+  units_options(simplify = FALSE)
+  expect_equal(set_units(1, "g/g"), x_gg)
+  expect_equal(set_units(1, "g") / set_units(1, "g"), x_gg)
+
+  units_options(simplify = TRUE)
+  expect_equal(set_units(1, "g/g"), x_ul)
+  expect_equal(set_units(1, "g") / set_units(1, "g"), x_ul)
+
+  units_options(simplify = NA)
+  expect_equal(set_units(1, "g/g"), x_gg)
+  expect_equal(set_units(1, "g") / set_units(1, "g"), x_ul)
+})